Move cmd/dockerd/debug to daemon/command/debug
