+++ daemon/command/debug/debug_test.go
