+++ daemon/info.go
@@ -17,7 +17,7 @@ import (
 	"github.com/docker/docker/api"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/system"
-	"github.com/docker/docker/cmd/dockerd/debug"
+	"github.com/docker/docker/daemon/command/debug"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/internal/filedescriptors"
 	"github.com/docker/docker/daemon/logger"
